+++ src/3rd_party_bundle_add.c
@@ -19,8 +19,6 @@
 
 #define _GNU_SOURCE
 
-#include <fcntl.h>
-
 #include "3rd_party_repos.h"
 #include "swupd.h"
 
@@ -28,11 +26,6 @@
 
 #define FLAG_SKIP_OPTIONAL 2000
 #define FLAG_SKIP_DISKSPACE_CHECK 2001
-
-#define SCRIPT_TEMPLATE "#!/bin/bash\n\n"                                \
-			"export PATH=%s:$PATH\n"                         \
-			"export LD_LIBRARY_PATH=%s:$LD_LIBRARY_PATH\n\n" \
-			"%s \"$@\"\n"
 
 static char **cmdline_bundles;
 static char *cmdline_repo = NULL;
@@ -101,86 +94,18 @@ static bool parse_options(int argc, char **argv)
 	return true;
 }
 
-static enum swupd_code create_wrapper_script(char *filename)
-{
-	enum swupd_code ret_code = 0;
-	int fd;
-	FILE *fp = NULL;
-	char *bin_directory = NULL;
-	char *script = NULL;
-	char *binary = NULL;
-	char *third_party_bin_path = NULL;
-	char *third_party_ld_path = NULL;
-	mode_t mode = 0755;
-
-	bin_directory = third_party_get_bin_dir();
-	script = third_party_get_binary_path(sys_basename(filename));
-	binary = sys_path_join(globals.path_prefix, filename);
-
-	if (!sys_filelink_is_executable(binary)) {
-		warn("File %s does not have 'execute' permission so it won't be exported\n", filename);
-		goto close_and_exit;
-	}
-
-	if (sys_file_exists(script)) {
-		/* the binary already exists, this condition should never happen
-		 * since we checked before installing */
-		ret_code = SWUPD_UNEXPECTED_CONDITION;
-		error("There is already a binary called %s in %s, it will be skipped\n", sys_basename(filename), bin_directory);
-		goto close_and_exit;
-	}
-
-	/* if the SWUPD_3RD_PARTY_BIN_DIR does not exist, attempt to create it */
-	if (mkdir_p(bin_directory)) {
-		ret_code = SWUPD_COULDNT_CREATE_DIR;
-		goto close_and_exit;
-	}
-
-	if (!is_dir(bin_directory)) {
-		error("The path %s for 3rd-party content exists but is not a directory\n", bin_directory);
-		ret_code = SWUPD_COULDNT_CREATE_DIR;
-		goto close_and_exit;
-	}
-
-	/* open the file with mode set to 0755 */
-	fd = open(script, O_RDWR | O_CREAT, mode);
-	if (fd < 0) {
-		error("The file %s failed to be created\n", script);
-		ret_code = SWUPD_COULDNT_CREATE_FILE;
-		goto close_and_exit;
-	}
-	fp = fdopen(fd, "w");
-	if (!fp) {
-		error("The file %s failed to be created\n", script);
-		ret_code = SWUPD_COULDNT_CREATE_FILE;
-		goto close_and_exit;
-	}
-
-	/* get the path for the 3rd-party content */
-	third_party_bin_path = str_or_die("%sbin:%susr/bin:%susr/local/bin", globals.path_prefix, globals.path_prefix, globals.path_prefix);
-	third_party_ld_path = str_or_die("%susr/lib64:%susr/local/lib64", globals.path_prefix, globals.path_prefix);
-
-	fprintf(fp, SCRIPT_TEMPLATE, third_party_bin_path, third_party_ld_path, binary);
-
-close_and_exit:
-	if (fp) {
-		fclose(fp);
-	}
-	free_string(&binary);
-	free_string(&script);
-	free_string(&bin_directory);
-	free_string(&third_party_bin_path);
-	free_string(&third_party_ld_path);
-
-	return ret_code;
-}
-
-static enum swupd_code validate_binary(char *filename)
+static enum swupd_code validate_binary(struct file *file)
 {
 	enum swupd_code ret_code = SWUPD_OK;
 	char *bin_directory = NULL;
 	char *script = NULL;
+	char *filename = NULL;
 
+	if (!file) {
+		return ret_code;
+	}
+
+	filename = file->filename;
 	bin_directory = third_party_get_bin_dir();
 	script = third_party_get_binary_path(sys_basename(filename));
 
@@ -197,7 +122,7 @@ static enum swupd_code validate_binary(char *filename)
 
 static enum swupd_code export_bundle_binaries(struct list *installed_files)
 {
-	return third_party_process_binaries(installed_files, "\nExporting 3rd-party bundle binaries...\n", "export_binaries", create_wrapper_script);
+	return third_party_process_binaries(installed_files, "\nExporting 3rd-party bundle binaries...\n", "export_binaries", third_party_create_wrapper_script);
 }
 
 static enum swupd_code validate_bundle_binaries(struct list *files_to_be_installed)
